requirements.txt
ShopifyAPI
dotenv